# src/openApi/reservation-api.yml
openapi: 3.0.3
info:
  title: HH3Week API
  description: >
    이 API는 사용자가 서비스를 이용하기 위해 토큰을 발급받고, 해당 토큰의 대기 번호를 조회할 수 있으며,
    예약 가능한 날짜와 해당 날짜의 좌석 정보를 조회할 수 있고, 잔액을 충전 및 조회하며, 결제를 처리하는 기능을 제공합니다.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
paths:
...
  /api/v1/payments/{paymentId}:
    get:
      summary: 결제 내역 조회
      description: >
        결제 ID를 통해 해당 결제의 상세 내역을 조회합니다.
      tags:
        - Payment
      parameters:
        - name: paymentId
          in: path
          description: 조회할 결제 ID
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      responses:
        '200':
          description: "결제 내역 조회 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentHistoryDto'
              example:
                paymentId: 10
                userId: 1
                reservationId: 2
                amount: 10000.0
                paymentStatus: "COMPLETED"
        '404':
          description: "결제 내역을 찾을 수 없음"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: "오류 유형"
                    example: "Not Found"
                  message:
                    type: string
                    description: "오류 상세 메시지"
                    example: "결제 내역을 찾을 수 없습니다."
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: "오류 유형"
                    example: "Internal Server Error"
                  message:
                    type: string
                    description: "오류 상세 메시지"
                    example: "결제 내역 조회 중 오류가 발생했습니다."
components:
  schemas:
...
    PaymentHistoryDto:
      type: object
      properties:
        paymentId:
          type: integer
          format: int64
          description: "결제 ID"
          example: 10
        userId:
          type: integer
          format: int64
          description: "사용자 ID"
          example: 1
        reservationId:
          type: integer
          format: int64
          description: "예약 ID"
          example: 2
        amount:
          type: number
          format: double
          description: "결제 금액"
          example: 10000.0
        paymentStatus:
          type: string
          description: "결제 상태"
          enum: ["PENDING", "COMPLETED", "FAILED"]
          example: "COMPLETED"
      required:
        - paymentId
        - userId
        - reservationId
        - amount
        - paymentStatus
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...